+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * add_var_random_01() creates variable of type integer
 * add target_name & factorise_target parameter to more create_data_*()
+* add checks to create_data_*()
 
 ## explore 1.0.1 (2022-12-20)